+++ CHANGELOG.md
@@ -7,9 +7,15 @@
 * Removed @deprecated stuff:
   * **[edison-core]** Removed ServiceSpec. Replaced by ServiceDependency.
   * **[edison-cache]** Removed CacheRegistry. Caches can be exposed as Spring Beans instead.
-  * **[edison-mongo]** Default constructor from AbstractMongoRepository is replaced with AbstractMongoRepository(MongoProperties)
-  * **[edison-mongo]** Removed property `passwd` from MongoProperties which was replaced with `password` in order to prevent exposure
-  of passwords in /internal/env etc.
+  * **[edison-mongo]** Default constructor from AbstractMongoRepository is replaced with 
+  AbstractMongoRepository(MongoProperties)
+  * **[edison-mongo]** Removed property `passwd` from MongoProperties which was replaced with `password` in order to 
+  prevent exposure
+  * **[edison-mongo]** Removed property `socketTimeout` from MongoProperties. Use `defaultReadTimeout` and 
+  `defaultWriteTimeout` instead.
+  * **[edison-mongo]** Removed property `socketTimeoutForHighTimeoutClient` from MongoProperties, as well as Spring 
+  Beans `mongoClientWithHighSocketTimeout` and `mongoDatabaseWithHighSocketTimeout`. Use custom timeouts 
+  on read and write operations instead. 
   * **[edison-togglz]** LDAP support from edison-togglz is replaced by edison-core
 
 ## 1.2.3